+++ modules/carto/src/sources/base-source.ts
@@ -10,7 +10,7 @@ import {
   Tilejson,
   TilejsonMapInstantiation
 } from './common';
-import {requestWithParameters} from './utils';
+import requestWithParameters from '../api/request-with-parameters';
 
 export async function baseSource<UrlParameters extends Record<string, string>>(
   endpoint: MapType,
